feat: convert @azure/ms-rest-nodeauth to identity
Convert usages of `@azure/ms-rest-nodeauth` and `adal-node` packages to
use the equivalent `@azure/identity` functions.

BREAKING CHANGE: This introduces breaking changes as it replaces the previously
used `adal-node` package for Azure authentication with the new
`@azure/identity` package. Because of difference in these two libraries
we had to remove support for the `msiEndpoint` option of
the `azure-active-directory-msi-vm` authentication method, and the
`msiEndpoint` and `msiSecret` options of the
`azure-active-directory-msi-app-service` authentication option.

There could also be other changes in authentication behavior, so be careful
when upgrading if you're using one of the Azure specific authentication methods.